fix: round 2 production hardening — lifetime safety, fused kernels, unwrap elimination
Implemented PRs (Codex-validated):
- PR-01: RDMA PostedOp<'a> lifetime safety — PhantomData prevents MR drop before completion
- PR-03: Timeout/retry constants → RdmaConfig + ExchangeConfig (configurable)
- PR-04: sendrecv recv-pre-post pattern (UC data loss prevention) + rank validation
- PR-05: MoE 5x unwrap() removal + input validation (indices/weights/token_data)
- PR-06: Core 6x Mutex::lock().unwrap() → KernelError::Internal
- PR-07: DType::UInt32 + embedding index dtype fix (Float32 placeholder removed)
- PR-10: NN fused kernels — Linear bias O(1) sync, Attention head concat O(1) sync,
         MoE token copy O(1) sync (was O(batch)/O(heads*seq)/O(seq))
- PR-11: Python global GPU context (OnceLock singleton, per-op device creation removed)
- PR-12: MoeForwardMetrics connected to forward(), warmup → Result, RDMA tests #[ignore]

Codex-deferred (profile first): PR-08 (Condvar), PR-09 (MR pool)
Codex-skipped: PR-02 (bindgen — macOS has no librdma headers)

Verification: 311 tests (290→311), 0 failures, 0 clippy warnings,
0 unwrap in moe_exchange, 0 Mutex unwrap in kernels, 0 unreachable in nn.

+++ crates/rmlx-nn/src/embedding.rs
@@ -98,13 +98,13 @@ impl Embedding {
             byte_size,
             MTLResourceOptions::StorageModeShared,
         );
-        // Wrap in Array with Float32 dtype — the gather shader reads as uint* so the
-        // dtype is irrelevant; we just need the correct buffer, shape, and element count.
+        // Wrap in Array with UInt32 dtype — the gather shader reads as uint* and
+        // UInt32 correctly describes the index data type.
         let indices_arr = Array::new(
             idx_buffer,
             vec![seq_len * embed_dim],
             vec![1],
-            DType::Float32, // placeholder dtype — gather reads buffer(1) as uint*
+            DType::UInt32,
             0,
         );
